add: let user decide name/rev of release

+++ release.sh
@@ -1,10 +1,26 @@
 #/bin/sh
-NAME=peex-1.0
+LATEST=`git tag -l release-* | cut -b 9- | sort -n -r | head -n 1`
+echo -n "Enter release to build [$LATEST]? "
+read REL
+if [ -z $REL ]; then
+	REL="$LATEST"
+fi
 
-echo "Building $NAME"
+NAME="peex-$REL"
+if [ -d $NAME ]; then
+	echo -n "A directory called '$NAME' already exists. Remove? [N/y] "
+	read CONFIRM
+	if [ $CONFIRM != "y" ]; then
+		echo "Release directory exists, aborting."
+		exit 1
+	fi
+fi
+
+echo "Building release '$NAME'"
 rm -rf $NAME
 mkdir -p $NAME
 cp peex defaults.peex ops.py tree.py $NAME/
 cp README example.peex $NAME/
-tar -cf $NAME.tar $NAME/
-gzip $NAME.tar
+tar -cf $NAME/$NAME.tar $NAME/
+gzip $NAME/$NAME.tar
+
